+++ KNXEngine/protocol/KNXConnectResponse.js
@@ -1,13 +1,13 @@
 'use strict';
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.KNXConnectResponse = void 0;
-const KNXConstants_1 = require("./KNXConstants");
-const KNXPacket_1 = require("./KNXPacket");
-const HPAI_1 = require("./HPAI");
-const CRD_1 = require("./CRD");
-class KNXConnectResponse extends KNXPacket_1.KNXPacket {
+const KNXConstants = require("./KNXConstants");
+const KNXPacket = require("./KNXPacket");
+const HPAI = require("./HPAI");
+const CRD = require("./CRD");
+class KNXConnectResponse extends KNXPacket.KNXPacket {
     constructor(channelID, status, hpai, crd) {
-        super(KNXConstants_1.KNX_CONSTANTS.CONNECT_RESPONSE, hpai == null ? 2 : 2 + hpai.length + crd.length);
+        super(KNXConstants.KNX_CONSTANTS.CONNECT_RESPONSE, hpai == null ? 2 : 2 + hpai.length + crd.length);
         this.channelID = channelID;
         this.status = status;
         this.hpai = hpai;
@@ -21,27 +21,27 @@ class KNXConnectResponse extends KNXPacket_1.KNXPacket {
         const status = buffer.readUInt8(offset++);
         let hpai, crd;
         if (offset < buffer.length) {
-            hpai = HPAI_1.HPAI.createFromBuffer(buffer, offset);
+            hpai = HPAI.HPAI.createFromBuffer(buffer, offset);
             offset += hpai.length;
-            crd = CRD_1.CRD.createFromBuffer(buffer, offset);
+            crd = CRD.CRD.createFromBuffer(buffer, offset);
         }
         return new KNXConnectResponse(channelID, status, hpai, crd);
     }
     static statusToString(status) {
         switch (status) {
-            case KNXConstants_1.KNX_CONSTANTS.E_SEQUENCE_NUMBER:
+            case KNXConstants.KNX_CONSTANTS.E_SEQUENCE_NUMBER:
                 return 'Invalid Sequence Number';
-            case KNXConstants_1.KNX_CONSTANTS.E_CONNECTION_TYPE:
+            case KNXConstants.KNX_CONSTANTS.E_CONNECTION_TYPE:
                 return 'Invalid Connection Type';
-            case KNXConstants_1.KNX_CONSTANTS.E_CONNECTION_OPTION:
+            case KNXConstants.KNX_CONSTANTS.E_CONNECTION_OPTION:
                 return 'Invalid Connection Option';
-            case KNXConstants_1.KNX_CONSTANTS.E_NO_MORE_CONNECTIONS:
+            case KNXConstants.KNX_CONSTANTS.E_NO_MORE_CONNECTIONS:
                 return 'No More Connections';
-            case KNXConstants_1.KNX_CONSTANTS.E_DATA_CONNECTION:
+            case KNXConstants.KNX_CONSTANTS.E_DATA_CONNECTION:
                 return 'Invalid Data Connection';
-            case KNXConstants_1.KNX_CONSTANTS.E_KNX_CONNECTION:
+            case KNXConstants.KNX_CONSTANTS.E_KNX_CONNECTION:
                 return 'Invalid KNX Connection';
-            case KNXConstants_1.KNX_CONSTANTS.E_TUNNELING_LAYER:
+            case KNXConstants.KNX_CONSTANTS.E_TUNNELING_LAYER:
                 return 'Invalid Tunneling Layer';
             default:
                 return `Unknown error ${status}`;